Create initial data model and rules engine
Implements a broad-brush view of the S&L data model, and a simple
encapsulation of Drools as a rules engine, with some rules to do
some logging and a single rule to create problems when a session
is booked with a non-available participant.

A test class demonstrates how the rule creates problems and how
the problems are removed when the input facts that they depend on
are changed.

Note that this builds with Maven - the pom.xml is included to
make the dependencies clear, but the dependencies will need to
be added to the gradle build before merging.

Note also that there are some vestigial dependencies from a version
of the code which also had some JPA persistence implemented.