Get parallel rules engines working
This version runs two rules engines, A and B, with rules in each which
create facts the other is interested in.  It demonstrates that facts can
be routed between engines.  Currently it creates facts uncommitted (i.e.
with Asserting facts present).  The next stage is to implement commit
and rollback functionality.